requirements.txt
pandas
numpy
python-dotenv
ShopifyAPI
